requirements.txt
streamlit>=1.33.0
sqlmodel>=0.0.16
openai>=1.13.3
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
backoff>=2.2.0
urllib3>=2.0.0
certifi>=2024.12.0
requests>=2.31.0
distro==1.9.0
docutils==0.21.2
dotty-dict==1.3.1
entrypoints==0.4
et_xmlfile==2.0.0
Faker==36.1.1
favicon==0.7.0
fonttools==4.54.1
frozenlist==1.6.2
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.12.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==7.2.1
iniconfig==2.1.0
invoke==1.7.3
japanize-matplotlib==1.1.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kiwisolver==1.4.7
lightgbm==4.5.0
lxml==5.3.1
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.4.4
narwhals==1.19.1
nh3==0.2.20
numpy==1.26.4
openai==1.86.0
openpyxl==3.1.5
packaging==23.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
pkginfo==1.12.1.2
plotly==5.18.0
pluggy==1.6.0
postgrest==1.0.2
prometheus_client==0.21.1
propcache==0.3.1
protobuf==4.25.6
psycopg2-binary==2.9.9
pyarrow==18.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pymdown-extensions==10.14.3
pyparsing==3.2.0
pytesseract==0.3.13
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-gitlab==3.15.0
python-semantic-release==7.33.2
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
realtime==2.4.3
referencing==0.35.1
reportlab==4.0.7
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
scipy==1.14.1
seaborn==0.13.2
semver==2.13.0
setuptools==78.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlmodel==0.0.24
st-annotated-text==4.0.2
st-theme==1.2.3
storage3==0.11.3
streamlit==1.45.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
twine==3.8.0
typing_extensions==4.14.0
tzdata==2024.2
tzlocal==5.3.1
urllib3==2.4.0
uuid==1.30
validators==0.34.0
watchdog==6.0.0
websockets==12.0
wheel==0.45.1
yarl==1.20.0
zipp==3.21.0
yfinance==0.2.38
PyPDF2==3.0.1
pdfplumber==0.11.4
python-multipart==0.0.9
